logout

+++ controllers/contacts.js
@@ -1,35 +1,54 @@
 const { Contact } = require('../models/contact');
 
 async function listContacts(req, res, next) {
-  const contacts = await Contact.find({});
+  const user = req.user._id;
+
+  const contacts = await Contact.find({ owner: user });
   res.status(200).json(contacts);
 }
 
 async function getContactById(req, res, next) {
   const { contactId } = req.params;
+  const user = req.user._id;
+
   const contact = await Contact.findById(contactId);
+
   if (!contact) {
-    return res.status(404).json({ message: 'Not found' });
+    return res.status(404).json({ message: 'contact not found' });
   }
+
+  if (!contact.owner.equals(user)) {
+    return res.status(401).json({ message: 'not authorize' });
+  }
+
   res.status(200).json(contact);
 }
 
 const removeContact = async (req, res, next) => {
   const { contactId } = req.params;
-  const contact = await Contact.findByIdAndRemove(contactId);
+  const user = req.user._id;
+
+  const contact = await Contact.findById(contactId);
+
   if (!contact) {
-    return res.status(404).json({ message: 'Not found' });
+    return res.status(404).json({ message: 'contact not found' });
   }
+
+  if (!contact.owner.equals(user)) {
+    return res.status(401).json({ message: 'not authorize' });
+  }
+
+  await Contact.findByIdAndRemove(contactId);
   res.status(200).json({ message: `contact "${contact.name}" deleted` });
 };
 
 const addContact = async (req, res, next) => {
   try {
-    const newContact = await Contact.create(req.body);
+    const newContact = await Contact.create({ ...req.body, owner: req.user._id });
     res.status(201).json(newContact);
   } catch (error) {
     if (error.code === 11000) {
-      return res.status(400).json({ message: 'Name must be unique' });
+      return res.status(400).json({ message: `Contact "${req.body.name}" already exist` });
     }
     res.status(400).json({ message: error.message });
   }
@@ -38,18 +57,27 @@ const addContact = async (req, res, next) => {
 const updateContact = async (req, res, next) => {
   try {
     const { contactId } = req.params;
+    const user = req.user._id;
+
+    const contact = await Contact.findById(contactId);
+
+    if (!contact) {
+      return res.status(404).json({ message: 'contact not found' });
+    }
+
+    if (!contact.owner.equals(user)) {
+      return res.status(401).json({ message: 'not authorize' });
+    }
+
     const updatedContact = await Contact.findByIdAndUpdate(
       contactId,
       { ...req.body },
       { new: true }
     );
-    if (!updatedContact) {
-      return res.status(404).json({ message: 'Not found' });
-    }
     res.status(200).json(updatedContact);
   } catch (error) {
     if (error.code === 11000) {
-      return res.status(400).json({ message: `Name "${req.body.name}" already exist` });
+      return res.status(400).json({ message: `Contact "${req.body.name}" already exist` });
     }
     res.status(400).json({ message: error.message });
   }
@@ -57,14 +85,22 @@ const updateContact = async (req, res, next) => {
 
 const updateContactStatus = async (req, res, next) => {
   const { contactId } = req.params;
-  const updatedContact = await Contact.findByIdAndUpdate(
-    contactId,
-    { favorite: req.body.favorite },
-    { new: true }
-  );
-  if (!updatedContact) {
-    res.status(404).json({ message: 'Not found' });
+  const user = req.user._id;
+
+  const contact = await Contact.findById(contactId);
+
+  if (!contact) {
+    return res.status(404).json({ message: 'contact not found' });
+  }
+
+  if (!contact.owner.equals(user)) {
+    return res.status(401).json({ message: 'not authorize' });
   } else {
+    const updatedContact = await Contact.findByIdAndUpdate(
+      contactId,
+      { favorite: req.body.favorite },
+      { new: true }
+    );
     res.status(200).json(updatedContact);
   }
 };
